+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 ## In development
 
 - Consolidate lesson definition logic, and provide local validation tool (syrctl) [#30](https://github.com/nre-learning/syringe/pull/30)
-
+- Redesign and fix the way iframe resources are created and presented to the API[#32](https://github.com/nre-learning/syringe/pull/32)
 
 ## 0.1.3 - November 15, 2018